Add files via upload
added in edit note page with ability to edit existing notes and the notes update based on the user changes. Also the user is directed to the edit note page when creating a note. The edit note page has a remove note button, and a return to home page button.